# backend-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.FileTypeRule:
    properties:
      createdAt:
        type: string
      description:
        type: string
      extension:
        type: string
      fileType:
        type: string
      id:
        type: integer
      status:
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.LoginDevice:
    properties:
      browser:
        type: string
      createdAt:
        type: string
      deviceId:
        type: string
      deviceName:
        type: string
      deviceType:
        type: string
      id:
        type: integer
      ip:
        type: string
      isCurrent:
        type: integer
      lastActiveAt:
        type: string
      location:
        type: string
      os:
        type: string
      userId:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 后端管理服务 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Server API
  version: "1.0"
paths:
...
  /auth/devices:
    get:
      description: 获取当前用户所有已登录设备
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.LoginDevice'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取登录设备列表
      tags:
      - 登录设备
  /auth/devices/{id}:
    delete:
      description: 将指定设备踢下线
      parameters:
      - description: 设备 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 踢出指定设备
      tags:
      - 登录设备
  /auth/devices/kick-all:
    delete:
      description: 踢出除当前设备外的所有设备
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    format: int64
                    type: integer
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 踢出所有其他设备
      tags:
      - 登录设备
...
  /system/file-type-rules:
    get:
      description: 获取所有文件类型规则
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.FileTypeRule'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取文件类型规则列表
      tags:
      - 文件类型规则
    post:
      consumes:
      - application/json
      description: 创建新的文件类型规则
      parameters:
      - description: 文件类型规则
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.FileTypeRule'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.FileTypeRule'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建文件类型规则
      tags:
      - 文件类型规则
  /system/file-type-rules/{id}:
    delete:
      description: 删除指定文件类型规则
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除文件类型规则
      tags:
      - 文件类型规则
    put:
      consumes:
      - application/json
      description: 更新指定文件类型规则
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文件类型规则
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.FileTypeRule'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.FileTypeRule'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新文件类型规则
      tags:
      - 文件类型规则
  /system/file-type-rules/grouped:
    get:
      description: 获取按文件类型分组的规则
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取分组文件类型规则
      tags:
      - 文件类型规则
  /system/file-type-rules/refresh:
    post:
      description: 从数据库刷新 AutoTagger 的文件类型规则
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 刷新自动标签器规则
      tags:
      - 文件类型规则
...
securityDefinitions:
  BearerAuth:
    description: JWT Token（格式：Bearer {token}）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"